クライアントAPIの旧ルート削除
