+++ app/api/DB/host.ts
@@ -613,6 +613,7 @@ export class MongoDBHost implements DB {
   }
 
   async createEncryptedMessage(data: {
+    roomId: string;
     from: string;
     to: string[];
     content: string;
@@ -620,6 +621,7 @@ export class MongoDBHost implements DB {
     encoding?: string;
   }) {
     const doc = await HostEncryptedMessage.create({
+      roomId: data.roomId,
       from: data.from,
       to: data.to,
       content: data.content,
@@ -740,11 +742,13 @@ export class MongoDBHost implements DB {
   }
 
   async createHandshakeMessage(data: {
+    roomId: string;
     sender: string;
     recipients: string[];
     message: string;
   }) {
     const doc = new HostHandshakeMessage({
+      roomId: data.roomId,
       sender: data.sender,
       recipients: data.recipients,
       message: data.message,
